Update iOS and Android
