+++ ios/phone_number.podspec
@@ -1,9 +1,10 @@
 #
-# To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html
+# To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html.
+# Run `pod lib lint phone_number.podspec' to validate before publishing.
 #
 Pod::Spec.new do |s|
   s.name             = 'phone_number'
-  s.version          = '0.6.2'
+  s.version          = '0.8.0'
   s.summary          = 'Flutter plugin for phone number validation'
   s.description      = <<-DESC
 Flutter plugin for phone number validation
@@ -13,11 +14,12 @@ Flutter plugin for phone number validation
   s.author           = { 'Your Company' => 'email@example.com' }
   s.source           = { :path => '.' }
   s.source_files = 'Classes/**/*'
-  s.public_header_files = 'Classes/**/*.h'
   s.dependency 'Flutter'
-  s.dependency 'PhoneNumberKit/PhoneNumberKitCore', '~> 3.2.0'
-  s.swift_versions = '5.0'
-  
-  s.ios.deployment_target = '8.0'
-  s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES', 'VALID_ARCHS[sdk=iphonesimulator*]' => 'x86_64' }
+  s.platform = :ios, '9.0'
+
+  s.dependency 'PhoneNumberKit/PhoneNumberKitCore', '~> 3.3'
+
+  # Flutter.framework does not contain a i386 slice.
+  s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES', 'EXCLUDED_ARCHS[sdk=iphonesimulator*]' => 'i386' }
+  s.swift_version = '5.0'
 end
